Initial commit. Basic processing architecture.
A basic architecture for the synchronizer has been laid out. The
GUI is based on wxWidgets and uses wxGLcanvas to interact with
OpenCV through an OpenGL canvas. A simple 24Hz timer is used to
keep the processing rhythm.

The system inputs consist of a PDF file containing a presentation
and a footage video file containing a recording of such presentation.

The system outputs consist of a synchronization file, which lists
a sequence of timed instructions which reproduce what's seen on the
footage; and a video slideshow of the presentation, which follows
the synchronization instructions, ready for further video editing
and mixing. The processing to obtain such outputs has not been
implemented yet.

ImageMagick is used to read PDF files into suitable images.