Synchronization initialization.
The initialization step of the synchronization loop has been
implemented. It precomputes the slides' keypoints and descriptors
as well as matching the first frame with the first slide to find
the presentation projection in the footage, (which will be used
to speed up the tracking and matching of the presentation in the
next frames).

The types of some variables have been modified to align them with
related operations (so the amount of conversions are minimized,
improving performance and quality), e.g. int to double,
double to float.

The SyncLoop now has a feature detector and a matcher, as well as
cached keypoints and descriptors to avoid computing them twice.

BRISK has been chosen as a feature detector. In preliminary
experiments it shows superior performance with text, which are
common on slides. The matcher is based on the Hamming distance
and brute force, filtered by RANSAC and a maximum distance ratio
between the best match and the second best match.