WIP: Frame tracker general flow.
The main ideas for the tracker have been laid out. Nothing has been
tested yet and many bugs are expected, but at least the architecture
should be solid.

The system matches every frame with a reference frame using BRISK
features. If the match is good enough, the frame is discarded and
processing continues with the next one. If the match is bad, as
it is expected when the slide changes, neighbouring slides are
compared with the frame to see if any of them is a better match.

Note that even if the slide changes and there are no matches in
the presentation area of the footage, the rest of the frame could
still be enough for a good match, so the "goodness-of-fit test"
is performed considering only the keypoints inside the presentation
area.

If the camera has moved too much or the slide has changed, a new
reference frame is set.

Convex Quads can now test if a point is inside them. They have
augmented with edge normals to do so.