Completed first working version of tracker.
A lot of small errors, such as off-by-one index access, incorrect
assumptions about returned vector sizes, incorrect initializations,
missing corner cases, missing castings and mistyped equations,
have been fixed.

The tracker is now able to correctly follow the presentation, even
when the camera moves (only minor movements for now).

Next step is to generate keyframe when the camera moves too much
(so far, they are only generated when the slide changes). The
tracker can definitely keep up with some movement, unless the camera
man goes crazy (as in my test data) and believes it's a Hollywood
test of the zoom function. As long as the heavy movements do not
happen at the same time as the slide changes, or if there is enough
background to match independently of the slide, the tracker should
be fine.

Some debugging visual drawing has been included to observe the
intermediate results.

Some file-local functions have been promoted to their own module
to be used in the rest of the project.

Note that there are many corner cases to be solved, such as empty
slides and transition slides (which are slides very similar between
each other).