No latency render.
Until now, the rendering was being done using messages to the event
loop which is fine if the processing takes less time than the timer,
but it leads to heavy laggin if that's not the case.
Now, the canvas is redrawn directly without going through the event
loop. It looks smoother (although it does not seem to play nice
with the resize operation; I guess that operation is still trying
to redraw using the event loop, which is still blocked).

In the long term, it would be nice to use another thread or yielding
to accomodate the heavy processing with a smooth GUI, but this
suffices for the initial testing of the algorithms.