Motion keyframes.
Now, if the camera moves too much (and therefore the presentation
Quad moves or deforms too much), a new keyframe is inserted.

This helps with keeping both the drift and deviation errors low,
since keyframes are sparse in time, except when they're needed
(when the camera is moving). It is essential for a good tracking,
since otherwise we would be matching frames from very different
viewpoints (in particular, the presentation could be out-of-frame
in one of the shots, which is easier to track with intermediate
keyframes to observe the transition).