Tracking recovery.
The tracker can get lost sometimes, due to a single bad frame which
is able to pass the quality tests or because the camera man was
so crazy nothing made sense anymore.

A couple of system have been established to correct this problems.

First, in the hard frames, the whole frame is compared with the
slide images (in contrast with the previous behaviour which only
compared the slides with the "presentation region" in the frame,
specified by the slidepose Quad). Using the "presentation region"
gives cleaner results, but if the tracker goes wrong, then it becomes
counterproductive, since the region points to the wrong place
(usually a random and non-convex region). Taking the whole frame
is more robust, although the results are "dirtier" (i.e. a couple
of pixels off). A better solution could be to try to use the
presentation region first and only fall back to the full frame if
the region gives unimpressive results. Anyway, the current solution
works OK for this stage's objectives and tests.

Second, the Quad deviation and deformation costs are important to
obtain good performance but they mean that if the presentation Quad
gets stuck in a position too far from the real place, they will
artificially inflate the error for the real match and the system
will refuse to accept it because of such error.
This is expected since under normal circumstances such a unlikely
transition would not be acceptable. It's only legitimately observed
because the system got stuck (e.g. because of fast camera operation
 motion blur, automatic white balancing). So, the tracker now also
compares the frame with the previous one (in addition to the
reference frame). If the Quad is similar to the one in the previous
frame for three consecutive frames, it is deemed good enough and
accepted regardless of its cost. This fixes most of the sticking
problems.

Third, there are moments where the tracker gets really lost. This
usually happens at the end of the presentation, when the user runs
to a particular slide to answer a question. The tracker can jump
up to three slides per frame, assuming it can recognize the new
slides at the first try (i.e. no auto-exposure blinding), but even
that could be too slow for an avid presenter. So, if the tracker
is lost for seven frames in a row, it will check every single
slide instead of the only the current slide's neighbourhood.

With these modifications, the system is completely able to track
the test presentations (minus some difficult slides which do not
have good keypoints).