Immutable Quads.
The Quad class has been immutabilized. As it was, the class was
inconsistent because normals were only calculated on construction
but the vertices could be changed later, which would not update
such normals, leaving the object in an inconsistent state.

Mutability does not really make sense for this class, since quad
are seldomly modified (and when they are, a fresh new Quad is
required) and every possible modification basically boils down to
recomputing every single property, i.e. constructing a new Quad
from scratch internally.