Processing yielding.
Several Yield() calls have been added to the main processor
functions, so the GUI is more responsive.

Yield() calls continue processing other events in the queue. One of
these events could destroy the window mid-processing (e.g. if the
user clicks on Exit). So, the rest of the processing could segfault.
Fortunately, the only field in the timer that is tied to the window
hierarchy is the canvas, so now the main window notifies the timer
about its own destruction by stopping the timer and the timer
checks this IsRunning flag to avoid touching the canvas if it's
no longer valid.

The long-term solution is a separate thread, but for now this works
fine.

Some debug drawing has been commented out.

Additionally, some minor refactoring have been done to conform to
the "style guidelines".