Preparation for video generation.
After the synchronization was completed, the program would just
stop processing input and do nothing. Now it moves on to the second
stage of video generation (not implemented yet).

In the sync loop, every time the slide index is changed, an
instruction is written down. At the end of the synchronization,
this raw SyncInstructions objects is written to disk.

Processing loops now communicate with the rest of the world via
events. This way the app can stop the timers when their purpose
has been accomplished and swap them for new ones (in this case,
the synchronization is replaced by the video generation loop).

The SyncInstructions ToString() method had a typo that would
make the output incompatible with the corresponding string constructor
(it missed a space, which would be parsed incorrectly). Also, although
the slide index starts at zero internally, the human readable file
now starts slides at one.

The declarations of SlideSyncApp and SlideSyncWindow have been
moved to the header file.

Some functions have been moved to the util file.

The quadmask function was not used so it has been removed.

Other minor refactoring and documentation corrections.