Synchronization cache.
SyncLoop now checks the synchronization cache file before starting
to process the video. If it finds the file and is able to parse it,
it will bypass the entire loop and immediately raise the LoopFinished
event with the SyncInstructions from file.

The Skip() reading function (an overload of operator>> for streams)
incorrectly used a boolean flag as an IO manipulator. This has been
corrected.

The SyncInstructions string constructor has been generalized to accept
any stream, so there's no need for an in-memory cache of the
instructions, which can be read directly from file now.