Video generation.
After the program has synchronized the slides with the footage,
it now moves on to generate a slideshow video from such instructions.

A secondary slide array is extracted from the slides PDF file, but
at higher resolution and in color (8-8-8-bit RGB). This images are
used in the output video.

A new dependency has been introduced. For video encoding, the
project now uses the FFmpeg library.

The FFmpeg library has been hidden in the new VideoEncoder class,
which can receive frames and output them to a video file transparently.
It can handle any video format (deduced by the filename extension)
FFmpeg can handle, but the bitrate is fixed at 2 MiBs (which is enough
for a decent slideshow using H.264). Similarly the output format is
fixed at 1920x1080 (or smaller if the slides aspect ratio is different).
All these options may become tunable in the future.

The SyncInstructions class has been modified to use the (new)
SyncInstruction struct instead of tuples as its list elements. This
way, it is more readable and maintainable. Also, a new "End"
instruction has been added, which signals the stream has ended
(useful to mark the final timestamp, so the encoder can fill the
video with the last slide).

The new GenLoop processing loop is in charge of generating the video.
It is a light timer-based wrapper around the VideoEncoder class and
takes care of yielding and raising events.

Other minor refactorizations and style corrections.

FFmpeg is exhibiting a bug at the moment where it reads the framerate
incorrectly if the format is MP4/H.264. So, although common framerates
are 23.976 and 29.97, only integer framerates will be accepted
(otherwise, they will be rounded to the nearest integer).